Consolidated directory structure
add requirements.txt
o
