+++ cool_gifs/templates/gif_store/image_list.html
