+++ cool_gifs/tests.py
