+++ cool_gifs/views.py
