+++ requirements.txt
@@ -1,0 +1,3 @@
+Django ~= 2.1
+Pillow ~= 6.0
+gunicorn ~= 19.9
